refactor lora_test_utils
